Fix the APK build: float cannot widen and box to Double
`PluginCall.getDouble(String, Double)` takes a boxed default, and Java will
widen a primitive or box it but never both in one conversion, so passing the
`float` field straight in does not compile. Cast it at the call site.

This should have been caught before the push. It was not, because the
desktop stub used for typechecking declared the getters with primitive
parameters, where `float` widens to `double` happily. The stubs now carry the
real signatures, copied from @capacitor/android, and reproduce this exact
error against the unfixed line.

+++ native/nightlight/android/src/main/java/gr/nifo/nightlight/NightLightPlugin.java
@@ -150,7 +150,7 @@ public class NightLightPlugin extends Plugin {
         if (call.hasOption("dayKelvin")) c.dayKelvin = Curve.clampKelvin(call.getInt("dayKelvin", c.dayKelvin));
         if (call.hasOption("nightKelvin")) c.nightKelvin = Curve.clampKelvin(call.getInt("nightKelvin", c.nightKelvin));
         if (call.hasOption("transitionMin")) c.transitionMin = Math.max(1, Math.min(240, call.getInt("transitionMin", c.transitionMin)));
-        if (call.hasOption("intensity")) c.intensity = (float) Math.max(0, Math.min(1, call.getDouble("intensity", c.intensity)));
+        if (call.hasOption("intensity")) c.intensity = (float) Math.max(0, Math.min(1, call.getDouble("intensity", (double) c.intensity)));
 
         int step = Math.max(5, Math.min(60, call.getInt("step", 15)));
         JSArray out = new JSArray();
